Fix operator-streaming CI failures on MKL backends
Two operator-streaming tests passed on macOS (Accelerate) but failed under
Intel MKL on CI, both platform-numerics issues:

- _stream_operator_eval accumulated the field mean via a float32 ub.sum(),
  which drifts on MKL. On a (near-)constant target this kept the streamed
  ss_tot spuriously non-zero while the resident float64 mean drove it to the
  1e-12 guard, giving a ~100x R2 divergence. Accumulate the mean and the
  mean-subtracted ss_tot in float64 so ss_tot collapses to exactly 0 on any
  backend (deterministic IEEE float64), matching the resident field-R2. The
  constant-target test now uses a loose tolerance (it is a degenerate,
  R2-undefined regime; the invariant is only that the two do not diverge by
  orders of magnitude).

- The empty-streamed-operator-predict path forwarded a zero-length batch,
  and the operator FFT rejects a 0-size input on MKL. Return an empty
  (0, *grid) field directly without forwarding (the resident path forwards
  too, so it also crashed on MKL -- the "returns (0,*grid)" behavior was
  Accelerate-only). Test no longer compares against the resident forward.

Full unit suite green locally (282 passed).

+++ ilmarinen/core/allgraph.py
@@ -2605,14 +2605,17 @@ class AllGraph(_ContractFitMixin, _ReportsMixin, _PersistenceMixin, _SizeSelecti
         is left in TRAIN mode like the resident eval; operator primitives are per-sample (no batch-coupling), so
         the sequential arange chunking is score-neutral."""
         n = len(source)
-        # pass 1: global mean of the target fields
+        # pass 1: global mean of the target fields, accumulated in FLOAT64 (matches the resident
+        # uy.astype(np.float64).mean(); a float32 sum drifts platform-dependently -- e.g. under MKL -- and on a
+        # (near-)constant target that drift keeps ss_tot spuriously non-zero instead of collapsing to the guard).
         sum_u, count = 0.0, 0
         for j in range(0, n, bs):
             ub = source.u(np.arange(j, min(j + bs, n)))
-            sum_u += float(ub.sum().item())
+            sum_u += float(ub.double().sum().item())
             count += int(ub.numel())
         mean = sum_u / max(count, 1)
-        # pass 2: residual and (mean-subtracted) total sums of squares
+        # pass 2: residual (float32, matching the resident float32 (pred-u)^2 sum) and the mean-subtracted total
+        # sum of squares (float64, matching the resident float64 (u - float64 mean)^2 sum).
         ss_res, ss_tot = 0.0, 0.0
         for j in range(0, n, bs):
             ids = np.arange(j, min(j + bs, n))
@@ -2620,7 +2623,7 @@ class AllGraph(_ContractFitMixin, _ReportsMixin, _PersistenceMixin, _SizeSelecti
             with torch.no_grad():
                 pred = net(ab, xb).cpu()
             ss_res += float(((pred - ub) ** 2).sum().item())
-            ss_tot += float(((ub - mean) ** 2).sum().item())
+            ss_tot += float(((ub.double() - mean) ** 2).sum().item())
         return 1.0 - ss_res / (ss_tot + 1e-12)
 
     def _resident_subsample(self, data, cap=None):
